began working on get_time function, lots of work done, new enumerate variable implemented (still need to adjust previous code to use it) some other stuff happened but I forgor...

+++ Advanced_Mode.cpp
@@ -8,7 +8,7 @@
 using namespace std;
 
 // function for enabling advanced mode
-void Advanced_Mode(short int amt, char type)
+void Advanced_Mode(int amt, char type)
 {
     char toggle;
 
@@ -33,3 +33,4 @@ void Advanced_Mode(short int amt, char type)
     }
     
 }
+// changed stuff
